% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_kernel.R
\name{calc_dist_kernel}
\alias{calc_dist_kernel}
\title{Calculates the spatial kernel value}
\usage{
calc_dist_kernel(dist_mat, rasterl, alpha, p, aa)
}
\arguments{
\item{rasterl}{The RasterLayer object containing the population data}

\item{alpha}{The destination population power.}

\item{distance}{The distance between areas 1 and 2.}

\item{offset}{The offset distance.}

\item{gamma}{The distance power.}
}
\value{
Returns one matrix object containing the spatial kernel values.
}
\description{
Calculates the spatial kernel value between two areas using their distance, the destination
             population size, the destination population power, an offset distance and a distance power.
}
\examples{
K = calc_dist_kernel(x, y, 0.53, 10, 3)

}
