% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_kernel_shp.R
\name{calc_dist_kernel_shp}
\alias{calc_dist_kernel_shp}
\title{Calculates the spatial kernel value (Shapefile)}
\usage{
calc_dist_kernel_shp(dist_mat, dist_c, shp_data, alpha, p, p2, aa,
  delta = 0.3)
}
\arguments{
\item{dist_mat}{The distance matrix between all areas.}

\item{dist_c}{The threshold distance for the matched kernel. Note that setting this to an extremely high value
will effectively make the gravity kernel a smooth kernel with a single function since the threshold
distance will never be crossed.}

\item{shp_data}{The dataframe object containing the population data extracted from the Shapefile object
(generated by the \code{\link{prep_simulation_shp}} function)}

\item{alpha}{The destination population power.}

\item{p}{The distance power.}

\item{p2}{The secondary distance power for the matched kernel.}

\item{aa}{The offset distance.}

\item{delta}{The home cell modifier (to compensate the distance within cells equal to 0).}
}
\value{
Returns one matrix object containing the spatial kernel values.
}
\description{
Calculates the spatial kernel values between all areas in a Shapefile using their distance and
             population. This uses a gravity kernel structure (either matched or smooth depending on the threshold
             distance).
}
