% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toys.R
\name{create_toys}
\alias{create_toys}
\title{Generates low resolutions RasterLayer objects}
\usage{
create_toys()
}
\value{
Creates four RasterLayer objects in the Global Environment.
}
\description{
Generates low resolution versions of the Great Britain raster data to use when testing other
             functions. This is solely for rapid testing of other functions in the package. It is recommended
             to save these on the disk to fast load them again later using the load_toys function.
}
\examples{
create_toys()

}
