% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beta_shp.R
\name{find_beta_shp}
\alias{find_beta_shp}
\title{Fits beta from one resolution to another (Shapefile)}
\usage{
find_beta_shp(fit_data, shp, interval = 100, num_runs = 500)
}
\arguments{
\item{fit_data}{The incidence data to use for fitting.}

\item{interval}{The interval on which the fit should be performed, in days (default is 100 - i.e. the fit will
be performed on the first 100 days of the epidemic).}

\item{num_runs}{Number of stochastic runs to extract the median incidence for one beta value (default is 500).
Higher values will be longer to run, but yield more accurate results.}

\item{rasterl}{The Rasterlayer to fit on.}
}
\value{
Returns the fitted value of beta.
}
\description{
Adjusts beta to fit the incidence of one resolution onto another.
}
