% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pop_data.R
\name{load_pop_data}
\alias{load_pop_data}
\title{Loads required data for the simulation}
\usage{
load_pop_data()
}
\value{
Creates one RasterLayer object in the Global Environment.
}
\description{
Loads the UK total population density file from the working directory as a RasterLayer object.
}
\details{
This function is dangerous in the sense that it assigns a variable to the Global Environment
         automatically. If there are any errors, try running it with an empty Global Environment.
         Note that the function attempts to load the file in the RAM if possible for faster computation later.
}
\examples{
load_pop_data()
View(total_pop_data)

}
