% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raster.R
\name{load_raster}
\alias{load_raster}
\title{Loads required data for the simulation}
\usage{
load_raster()
}
\value{
Creates one RasterLayer object in the Global Environment.
}
\description{
Loads the total population density file of England as a RasterLayer object. The population is
             divided over a 100 metre grid.
}
\details{
Note that the function attempts to load the file in the RAM if possible for faster computation later.
}
\examples{

load_raster()
View(total_pop_data)

}
