% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shp.R
\name{load_shp}
\alias{load_shp}
\title{Loads all of the England Shapefiles, and the population data}
\usage{
load_shp()
}
\value{
Returns three Shapefile objects and one dataframe.
}
\description{
Loads three Shapefiles for administrative units of England (regions, counties and districts).
             Also loads the dataframe containing all of the population data for all the units.
}
\details{
This function assumes that the Shapefiles and population data are in specific directories. If it fails,
         manually load the shapefiles and population data using readShapeSpatial and read.csv functions.
}
\examples{

load_shp()

View(pop_data)

plot(region_shp)

}
