% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_toy.R
\name{load_toy}
\alias{load_toy}
\title{Loads a low resolution RasterLayer object}
\usage{
load_toy()
}
\value{
Loads one RasterLayer objects in the Global Environment.
}
\description{
Loads a low resolution version of the Great Britain raster data to use when testing other
             functions. This is solely for rapid testing of other functions in the package.
             Note that the function attempts to load the file in the RAM if possible for faster computation later.
}
\examples{
load_toy()
View(toy_data)

}
