% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_toys.R
\name{load_toys}
\alias{load_toys}
\title{Loads low resolutions RasterLayer objects}
\usage{
load_toys(in_memory = FALSE)
}
\arguments{
\item{in_memory}{Logical. If TRUE, the RasterLayer objects will be loaded in the RAM to avoid constantly
accessing the disk. These are small objects, therefore any system should be able to load them.
Nevertheless, this is set to FALSE as default for maximum compatibility.}
}
\value{
Loads four RasterLayer objects in the Global Environment.
}
\description{
Loads low resolution versions of the Great Britain raster data to use when testing other
             functions. This is solely for rapid testing of other functions in the package. These objects must
             have been created using the create_toys function.
}
\examples{
#does not load in RAM:
load_toys()
load_toys(in_memory = FALSE)

#loads in RAM:
load_toys(in_memory = TRUE)

}
