% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stoch_shp.R
\name{multi_stoch_shp}
\alias{multi_stoch_shp}
\title{Runs multiple stochastic simulations (Shapefile)}
\usage{
multi_stoch_shp(num_runs = 100, shp_data, D, contact_mat, beta,
  sigma = 1/2.6, step = 1, start_area = NA, start_num = 10,
  t_max = 200)
}
\arguments{
\item{num_runs}{The number of stochastic runs to perform.}

\item{shp_data}{The dataframe object containing the population data extracted from the Shapefile object
(generated by the \code{\link{prep_simulation_shp}} function)}

\item{D}{The expanded kernel matrix to use for FOI calculation (generated by the \code{\link{calc_beta_shp}}
function).}

\item{contact_mat}{The contact matrix for mixing between age groups.}

\item{beta}{The beta value for the epidemic (calculated from a given R0 using the \code{\link{calc_beta_shp}}
function).}

\item{sigma}{The recovery rate for the epidemic (must match the one used to calculate beta from R0
using the \code{\link{calc_beta_shp}} function).}

\item{step}{Size of time step for stochastic simulation, in days (default is 1 day).}

\item{start_area}{Where to start the epidemic. You can provide the name of the area directly (you can check
which are possible by looking inside the shp_data dataframe). If using the District, County
or Region resolution, the default is London. Otherwise, no default is provided.}

\item{start_num}{Number of infected individuals to start the epidemic.}

\item{t_max}{How many days to run the simulation for.}
}
\value{
Returns one dataframe object containing the median and 95% range of all of the simulation results.
}
\description{
Simulates multiple stochastic epidemics using the provided Shapefile, spatial kernel, contact
             matrix, and infection parameters.
}
\details{
This functions requires specific objects to run. These can be generated using the \code{\link{prep_simulation_shp}}
         function (e.g. if you want to simulate an epidemic using the Shapefile object "region_shp", you must
         prep_simulation_shp(region_shp) first). The model used is an SIR model, where individuals can be either
         Susceptible, Infected or Recovered with regards to the disease. This assumes that Infected individuals
         are infectious, and that Recovered individuals are immune and cannot be reinfected.
}
