% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_simulation.R
\name{prep_simulation}
\alias{prep_simulation}
\title{Performs pre-calculations necessary for the simulation of an epidemic}
\usage{
prep_simulation(rasterl, R0 = 1.8, sigma = 1/2.6)
}
\arguments{
\item{rasterl}{The RasterLayer object containing the population data.}

\item{R0}{The desired R0 for the epidemic}

\item{sigma}{The desired recovery rate for the epidemic}
}
\value{
Returns two matrix objects and one value for beta.
}
\description{
Creates the age mixing matrix and the spatial kernel using the load_contact_mat, calc_dist_mat,
             calc_dist_kernel and calc_beta functions.
}
\examples{
prep_simulation(toy_data, R0 = 2, sigma = 1/4.2)

}
