% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_simulation_shp.R
\name{prep_simulation_shp}
\alias{prep_simulation_shp}
\title{Performs pre-calculations necessary for the simulation of an epidemic (Shapefile)}
\usage{
prep_simulation_shp(shp, R0 = 1.8, sigma = 1/2.6)
}
\arguments{
\item{shp}{The Shapefile object containing the areas of interest}

\item{R0}{The desired R0 for the epidemic}

\item{sigma}{The desired recovery rate for the epidemic}
}
\value{
Returns two matrix objects and one value for beta.
}
\description{
Creates the age mixing matrix and the spatial kernel using the load_contact_mat, calc_dist_mat_shp,
             calc_dist_kernel_shp and calc_beta_shp functions.
}
\examples{
prep_simulation_shp(regions_shp, pop_data, R0 = 2, sigma = 1/4.2)

}
