% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_shp.R
\name{run_simulation_shp}
\alias{run_simulation_shp}
\title{Runs the spatial epidemic simulation (Shapefile)}
\usage{
run_simulation_shp(shp_data, dist_kernel, contact_mat, beta, alpha = 1,
  sigma = 1/2.6, stoch = FALSE, step = 1, start_area = NA,
  start_num = 1, t_max = 10)
}
\arguments{
\item{shp_data}{The dataframe object containing the population data extracted from the Shapefile object.}

\item{dist_kernel}{The distance kernel matrix.}

\item{contact_mat}{The contact matrix between age groups.}

\item{beta}{The beta value for the epidemic. (calculated from a given R0 using the calc_beta function)}

\item{sigma}{The desired recovery rate for the epidemic. (must match the one used to calculate beta from R0
using the calc_beta function)}

\item{stoch}{Logical. If TRUE, the simulation is stochastic.}

\item{step}{Size of time step for stochastic simulation (default is 1 day).}

\item{start_area}{Where to start the epidemic. You can provide the name of the area directly (you can check
which are possible by looking inside the pop_data dataframe). Default is area with highest
population (London).}

\item{start_num}{Number of infected individuals to start the epidemic.}

\item{t_max}{How many days to run the simulation for.}
}
\value{
Returns one dataframe object containing the epidemic estimates per day.
}
\description{
Simulates an epidemic using the provided Shapefile, spatial kernel, contact matrix, and
             infection parameters.
}
\details{
This functions requires specific objects to run. These can be generated using the prep_simulation_shp
         function (e.g. if you want to simulate an epidemic using the shp object "region_shp", you must run
         prep_simulation_shp(region_shp) first)
}
\examples{
results = run_simulation_shp(shp_data, kernel, contact_mat, beta = beta)

}
