% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_all_run.R
\name{test_all_run}
\alias{test_all_run}
\title{Quick testing function}
\usage{
test_all_run(R0 = 1.8, sigma = 1/2.6, t_max = 10,
  start_fraction = 1e-05)
}
\arguments{
\item{R0}{Desired R0 for the outbreak.}

\item{t_max}{Desired length of simulation.}

\item{start_fraction}{Desired starting proportion of infected in seeding area.}
}
\value{
Returns a summary of the sum of rows in the results (expected: only one value), the minimum value
        in the results (expected: 0) and the overall epidemic plot.
}
\description{
Performs setup (loading data, calculating distance kernel and contact matrix) and runs a short
             simulation to check that there are no errors in the calculations. Some customisation of parameters
             possible for extreme values testing.
}
\examples{
test_all_run(R0 = 2, sigma = 1/4.2, t_max = 100)

}
