% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_beta.R
\name{calc_beta}
\alias{calc_beta}
\title{Estimates beta from a given R0 (Raster)}
\usage{
calc_beta(rasterl, dist_kernel, contact_mat, R0 = 1.8, sigma = 1/2.6)
}
\arguments{
\item{rasterl}{The RasterLayer object containing the population data.}

\item{dist_kernel}{The distance kernel matrix.}

\item{contact_mat}{The contact matrix between age groups.}

\item{R0}{The desired value for R0.}

\item{sigma}{The desired value for the recovery rate.}
}
\value{
Returns the value of beta.
}
\description{
Estimates beta for an epidemic with a given R0 in a specific population in a Rasterlayer object.
}
\examples{
beta = calc_beta(toy_data, dist_kernel, contact_mat, R0=2, sigma=1/4.2)

}
