% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_kernel_shp.R
\name{calc_dist_kernel_shp}
\alias{calc_dist_kernel_shp}
\title{Calculates the spatial kernel value (Shapefiles)}
\usage{
calc_dist_kernel_shp(dist_mat, shp_data, alpha, p, aa, delta = 0.3)
}
\arguments{
\item{shp_data}{The Shapefile object containing the population data}

\item{alpha}{The destination population power.}

\item{distance}{The distance matrix between all areas.}

\item{offset}{The offset distance.}

\item{gamma}{The distance power.}
}
\value{
Returns one matrix object containing the spatial kernel values.
}
\description{
Calculates the spatial kernel value between all areas in a Shapefile using their distance, the
             destination population size, the destination population power, an offset distance and a distance
             power.
}
\examples{
K = calc_dist_kernel_shp(x, y, 0.53, 10, 3)

}
