% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_mat.R
\name{calc_dist_mat}
\alias{calc_dist_mat}
\title{Calculates the distance matrix (Raster)}
\usage{
calc_dist_mat(rasterl)
}
\arguments{
\item{raster_l}{The RasterLayer object to use to calculate the distance matrix.}
}
\value{
Returns one matrix object.
}
\description{
Calculates the distance between all populated areas in a RasterLayer object.
}
\details{
This function calculates the distance matrix once to avoid doing it for every generation during an
         epidemic simulation. This takes a while, but is essential to later speed up the epidemic simulation.
         The matrix returned only contains distance between cells with non-NA values. (i.e. cells with people
         in them)
}
\examples{
dist_mat = calc_dist_mat(total_pop_data)

}
