% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_mat_shp.R
\name{calc_dist_mat_shp}
\alias{calc_dist_mat_shp}
\title{Calculates the distance matrix (Shapefile)}
\usage{
calc_dist_mat_shp(shp)
}
\arguments{
\item{shp}{The Shapefile object to use to calculate the distance matrix.}
}
\value{
Returns one matrix object containg the distances between all populated areas.
}
\description{
Calculates the distance between all areas in a Shapefile object.
}
\details{
This function calculates the distance matrix once to avoid doing it for every generation during an
         epidemic simulation. This is essential to later speed up the epidemic simulation. Note that this
         function assumes that the distances are in meters in the dataset! If that is not the case, you might want
         to either convert these, or adjust the output of this function.
}
