% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pop_data.R
\name{load_pop_data}
\alias{load_pop_data}
\title{Loads required data for the simulation}
\usage{
load_pop_data()
}
\value{
Creates four RasterLayer objects and one matrix object in the Global Environment.
}
\description{
Loads the UK total population density raster file from the working directory, and divides it into
             4 objects corresponding to the four age groups. Also creates the age contact matrix.
}
\details{
This function is dangerous in the sense that it assigns variables to the Global Environment
         automatically. If there are any errors, try running it with an empty Global Environment.
         The approach of loading one raster file then subdividing it means that any person wishing to use this
         code only needs to have one .asc file in storage, compared to four.
         Note that this will work most effectively if your computer has at least 8Gb of RAM. Otherwise, the
         objects will not be loaded in memory and computation will be three times longer!
}
\examples{
load_pop_data()
View(contact_mat)
View(s0_4)

}
