% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Runs the spatial epidemic simulation}
\usage{
run_simulation(rasterl, dist_kernel, contact_mat, beta, sigma = 1/2.6,
  stoch = FALSE, start_area = 1, start_fraction = 0.01, t_max = 10)
}
\arguments{
\item{rasterl}{The RasterLayer object containing the population data.}

\item{dist_kernel}{The distance kernel matrix.}

\item{contact_mat}{The contact matrix between age groups.}

\item{beta}{The beta value for the epidemic. (calculated from a given R0 using the calc_beta function)}

\item{sigma}{The desired recovery rate for the epidemic. (must match the one used to calculate beta from R0
using the calc_beta function)}

\item{stoch}{Logical. If TRUE, the simulation is stochastic.}

\item{start_area}{Where to start the epidemic. 1: Most highly populated area (default), 2: A random area in
the middle of the country (typically medium population density), 3: A random area in the
north of the country (typically low population density).}

\item{start_fraction}{Fraction of infected individuals to start the epidemic.}

\item{t_max}{How many days to run the simulation for.}
}
\value{
Returns one dataframe object containing the epidemic estimates of cases per day.
}
\description{
Simulates an epidemic using the provided RasterLayers, distance matrix, spatial kernel,
             contact matrix, and infection parameters.
}
\details{
This functions requires many parameters to run. These can be generated using the following functions:
}
\examples{
results = run_simulation(total_pop_data, kernel, contact_mat, beta = beta)

}
