% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Runs the spatial epidemic simulation}
\usage{
run_simulation()
}
\arguments{
\item{raster_list}{A list containing the names of the four RasterLayers (one for each age category).}

\item{dist_matrix}{The distance matrix.}

\item{s_kernel}{The spatial kernel.}

\item{contact_mat}{The contact matrix between age groups.}

\item{R0}{The desired value for R0.}

\item{stoch}{Logical. If TRUE, the simulation is stochastic.}
}
\value{
Returns one dataframe object containing the total epidemic estimates of cases per day.
}
\description{
Simulates an epidemic using the provided RasterLayers, distance matrix, spatial kernel,
             contact matrix, and R0.
}
\details{
This functions requires many parameters to run. These can be generated using the following functions:
}
\examples{
dist_matrix = calc_dist_mat(s0_04)

}
